+++ src/test/scala/lktk/bmsg/structures/BloomFilterSpec.scala
@@ -1,0 +1,15 @@
+package lktk.bmsg.structures
+
+import lktk.bmsg.CodecSuite
+import lktk.bmsg.util.ProtocolVersion
+
+class BloomFilterSpec extends CodecSuite {
+
+  "BloomFilter codec" should {
+    "roundtrip" in {
+      bloomfilterGen.map { fl =>
+        roundtrip(BloomFilter.codec(ProtocolVersion.v70001), fl)
+      }.sample
+    }
+  }
+}
